+++ packages/health-check-nest/src/health-check/health-config.ts
@@ -3,6 +3,6 @@ export const HEALTH_CONFIG = Symbol('HEALTH_CONFIG');
 export type SupportedDatabase = 'postgres' | 'redis' | 'neo4j';
 
 export interface HealthConfig {
-  databases?: SupportedDatabase[];
-  failOnMissingProvider?: boolean;
+  databases: SupportedDatabase[];
+  failOnMissingProvider: boolean;
 }