+++ packages/health-check-nest/CHANGELOG.md
@@ -1,5 +1,15 @@
 # Changelog
 
+## 0.1.1
+
+### Patch Changes
+
+- Updated dependencies []:
+  - @volontariapp/errors-nest@0.5.0
+  - @volontariapp/bridge-nest@0.2.1
+  - @volontariapp/bridge@0.2.1
+  - @volontariapp/health-check@0.1.1
+
 All notable changes to this project will be documented in this file.
 
 ## 0.1.0
@@ -7,4 +17,3 @@ All notable changes to this project will be documented in this file.
 ### Minor Changes
 
 - Initial package scaffold.
-