+++ packages/health-check/CHANGELOG.md
@@ -1,5 +1,13 @@
 # Changelog
 
+## 0.1.1
+
+### Patch Changes
+
+- Updated dependencies []:
+  - @volontariapp/errors@0.4.0
+  - @volontariapp/bridge@0.2.1
+
 All notable changes to this project will be documented in this file.
 
 ## 0.1.0
@@ -7,4 +15,3 @@ All notable changes to this project will be documented in this file.
 ### Minor Changes
 
 - Initial package scaffold.
-